FAI-227 - Move category mapping to shared
