+++ src/SFA.DAS.SharedOuterApi/Interfaces/ICourseService.cs
@@ -1,0 +1,9 @@
+using System.Collections.Generic;
+
+namespace SFA.DAS.SharedOuterApi.Interfaces
+{
+    public interface ICourseService
+    {
+        List<string> MapRoutesToCategories(IReadOnlyList<string> routes);
+    }
+}
