+++ src/SFA.DAS.Vacancies.Api.UnitTests/AppStart/WhenAddingServicesToTheContainer.cs
@@ -25,6 +25,7 @@ namespace SFA.DAS.Vacancies.Api.UnitTests.AppStart
         [TestCase(typeof(IProviderRelationshipsApiClient<ProviderRelationshipsApiConfiguration>))]
         [TestCase(typeof(IAccountLegalEntityPermissionService))]
         [TestCase(typeof(IStandardsService))]
+        [TestCase(typeof(ICourseService))]
         public void Then_The_Dependencies_Are_Correctly_Resolved(Type toResolve)
         {
             var hostEnvironment = new Mock<IWebHostEnvironment>();
