+++ src/SFA.DAS.Vacancies/Application/Vacancies/Queries/GetVacanciesQueryHandler.cs
@@ -20,16 +20,19 @@ namespace SFA.DAS.Vacancies.Application.Vacancies.Queries
         private readonly IFindApprenticeshipApiClient<FindApprenticeshipApiConfiguration> _findApprenticeshipApiClient;
         private readonly IAccountLegalEntityPermissionService _accountLegalEntityPermissionService;
         private readonly IStandardsService _standardsService;
+        private readonly ICourseService _courseService;
         private readonly VacanciesConfiguration _vacanciesConfiguration;
 
         public GetVacanciesQueryHandler(IFindApprenticeshipApiClient<FindApprenticeshipApiConfiguration> findApprenticeshipApiClient, 
             IAccountLegalEntityPermissionService accountLegalEntityPermissionService, 
             IStandardsService standardsService,
+            ICourseService courseService,
             IOptions<VacanciesConfiguration> vacanciesConfiguration)
         {
             _findApprenticeshipApiClient = findApprenticeshipApiClient;
             _accountLegalEntityPermissionService = accountLegalEntityPermissionService;
             _standardsService = standardsService;
+            _courseService = courseService;
             _vacanciesConfiguration = vacanciesConfiguration.Value;
         }
 
@@ -59,7 +62,7 @@ namespace SFA.DAS.Vacancies.Application.Vacancies.Queries
                 }
             }
 
-            var categories = _standardsService.MapRoutesToCategories(request.Routes);
+            var categories = _courseService.MapRoutesToCategories(request.Routes);
 
             var vacanciesTask = _findApprenticeshipApiClient.Get<GetVacanciesResponse>(new GetVacanciesRequest(
                 request.PageNumber, request.PageSize, request.AccountLegalEntityPublicHashedId,
